Changed the way that disaggregation is handled. This did require me to alter the sql tables, so old databases need to be updated to include one new column in two of the tables.

+++ include/EventGraphWidget.h
@@ -274,7 +274,7 @@ private:
   QString selectedType;
   MacroEvent *selectedMacro;
   
-  int distance;
+  qreal distance;
   int vectorPos;
   int selectedIncident;
